try to solve shifting 32-bit by 251-bit

+++ projects/cm_mcu/LocalTasks.c
@@ -675,21 +675,21 @@ int8_t getFFpresentbit(const uint8_t i)
   configASSERT(i < NFIREFLIES);
   if (i < NFIREFLIES_IT_F1) {
     j = i;
-    val = (int8_t)(((ffl12_f1_args.present_bit_mask) >> j) & 0x01);
+    val = (int8_t)(((ffl12_f1_args.present_bit_mask) >> (int8_t)j) & 0x01);
   }
 
   else if (NFIREFLIES_IT_F1 <= i && i < NFIREFLIES_IT_F1 + NFIREFLIES_DAQ_F1) {
     j = i - NFIREFLIES_IT_F1;
-    val = (int8_t)(((ffldaq_f1_args.present_bit_mask) >> j) & 0x01);
+    val = (int8_t)(((ffldaq_f1_args.present_bit_mask) >> (int8_t)j) & 0x01);
   }
 
   else if (NFIREFLIES_F1 <= i && i < NFIREFLIES_F1 + NFIREFLIES_IT_F2) {
     j = i - NFIREFLIES_F1;
-    val = (int8_t)(((ffl12_f2_args.present_bit_mask) >> j) & 0x01);
+    val = (int8_t)(((ffl12_f2_args.present_bit_mask) >> (int8_t)j) & 0x01);
   }
   else {
     j = i - (NFIREFLIES_F1 + NFIREFLIES_IT_F2);
-    val = (int8_t)(((ffldaq_f2_args.present_bit_mask) >> j) & 0x01);
+    val = (int8_t)(((ffldaq_f2_args.present_bit_mask) >> (int8_t)j) & 0x01);
   }
 
   return val;
